rename compat to wrapper
