+++ src/cosmology/api/__init__.py
@@ -15,7 +15,7 @@ from cosmology.api._core import Cosmology
 from cosmology.api._extras import HasHubbleParameter, HasTcmb
 from cosmology.api._namespace import CosmologyNamespace
 from cosmology.api._standard import StandardCosmology
-from cosmology.api.compat import (
+from cosmology.api.wrapper import (
     CosmologyWrapper,
     StandardCosmologyWrapper,
 )
