+++ src/cosmology/api/wrapper/__init__.py
@@ -1,0 +1,9 @@
+"""The API standard for cosmology wrappers."""
+
+from cosmology.api.wrapper._core import CosmologyWrapper
+from cosmology.api.wrapper._standard import StandardCosmologyWrapper
+
+__all__ = [
+    "CosmologyWrapper",
+    "StandardCosmologyWrapper",
+]
